requirements/dev.txt
-r base.txt

# Development server
django-debug-toolbar>=4.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Code quality
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
django-stubs>=4.2.0

# Development utilities
ipython>=8.0.0
django-extensions>=3.2.0
watchfiles>=0.20.0